// angular-sample/angular-webapp/client/app/risk-assessment/models/risk-assessment-additionalcontrols.ts
import { ControlsCategory } from '../common/controls-category-enum';
export class RAAdditionalControl {
    Author: string;
    Category: number;
    Color: string;
    CompanyId: string;
    CreatedBy: string;
    CreatedOn: Date;
    Description: string;
    Graphic: string;
    Id: string;
    IsDeleted: false
    IsShared: false
    IsSharedPrototype: true
    LCid: number;
    ModifiedBy: string;
    ModifiedOn: string;
    Modifier: string;
    Name: string;
    PictureId: string;
    PrototypeId: string;
    RiskAssessment: string;
    RiskAssessmentId: string;
    Shape: string;
    Version: string;
    isSelected: boolean;
}